+++ src/lib.rs
@@ -27,6 +27,8 @@ pub mod internal_crypto;
 pub mod metrics;
 #[cfg(feature = "metrics-otlp")]
 pub mod metrics_otlp;
+#[cfg(feature = "metrics")]
+mod metrics_secret;
 #[cfg(all(feature = "web", feature = "proxy"))]
 mod native_http1_static;
 #[cfg(feature = "proxy")]